style: deleted extra file as subcomponent definition was moved to inside the main component
