+++ app/(pages)/(index-page)/hackdavis/_components/Sections/FinalSolution/FinalSolution.tsx
@@ -2,8 +2,6 @@
 import { motion } from 'framer-motion';
 import Image from 'next/image';
 import { useState } from 'react';
-import { CommentProps } from './Comment';
-import Comment from './Comment';
 import { useScrollToSectionOnViewSwitch } from '@/app/(pages)/_hooks/useScrollToSection';
 
 //TODO: comeback to this later to refactor again maybe ???
@@ -262,3 +260,25 @@ export const ToggleButton = ({
     </button>
   );
 };
+
+
+export type CommentProps = {
+  text: string;
+  x: string;
+  y: string;
+};
+
+export function Comment(props: CommentProps) {
+  const { text, x, y } = props;
+
+  return (
+    <div
+      className={`flex items-center absolute`}
+      style={{ right: `${x}`, top: `${y}` }}
+    >
+      <div className=" bg-[#005271] w-2 h-2 rounded-full"></div>
+      <div className="border-t-2 border-dashed border-[#005271] w-5"></div>
+      <p className="bg-white p-4 rounded-[4px] text-[14px] w-[192px]">{text}</p>
+    </div>
+  );
+}
